Rename creds rbac CRs for agent platform
"Cloud Credentials" for the Agent platform are given as RBAC CRs which enable the capi provider to watch and bind agents from a specific namespace. With current naming if the hostedControlPlane namespace happens to be the same than hostedCluster.Spec.Platform.Agent.AgentNamespace RBAC naming might clash.
This PR sets RBAC naming for Agent platform consistent with other references and add unit tests.