Reconcile AWS infra CR with NodePool subnets
There's an bug in CAPI AWS provider which prevent Machines from targeting subnets ids which are not present on the infra CR.
This PR is a workaround by keeping the infra CR subnets in sync with what is required by NodePools.
This should be dropped once https://github.com/kubernetes-sigs/cluster-api-provider-aws/pull/2864 gets merged.