fix replica param quoting in openshift-template generation
**What this PR does / why we need it**
The current template generation code wrongly quotes the int typed field
`.spec.replicas` of the operator deployment. the resulting template can not processesd.

`.spec.replicas` needs to be an int32 so having a quotes placeholder in
the openshift template is invalid. instead an the double curly bracket
syntax needs to be used to indicate that a value should be treated as
unquoted in the resulting manifests.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>